+++ ricaun.Revit.UI/RibbonItemExtension.cs
@@ -1,8 +1,5 @@
 ﻿using Autodesk.Revit.UI;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using UIFramework;
+using System.Windows.Media;
 
 namespace ricaun.Revit.UI
 {
@@ -11,246 +8,217 @@ namespace ricaun.Revit.UI
     /// </summary>
     public static class RibbonItemExtension
     {
+        #region Set RibbonItem
         /// <summary>
-        /// GetId from <paramref name="ribbonItem"/>
+        /// Sets the contextual help bound with this RibbonItem.
         /// </summary>
+        /// <typeparam name="TRibbonItem">RibbonItem</typeparam>
         /// <param name="ribbonItem"></param>
+        /// <param name="helpPath"></param>
         /// <returns></returns>
-        private static string GetId(this RibbonItem ribbonItem)
+        public static TRibbonItem SetContextualHelp<TRibbonItem>(this TRibbonItem ribbonItem, string helpPath) where TRibbonItem : RibbonItem
         {
-            var type = typeof(RibbonItem);
+            if (helpPath != null)
+                ribbonItem.SetContextualHelp(RibbonHelpExtension.GetContextualHelp(helpPath));
 
-            var parentId = type.GetField("m_parentId",
-                BindingFlags.Instance | BindingFlags.NonPublic)
-                ?.GetValue(ribbonItem)
-                ?? string.Empty;
-
-            var generateIdMethod = type.GetMethod("generateId",
-                BindingFlags.Static | BindingFlags.NonPublic);
-
-            return (string)generateIdMethod?.Invoke(ribbonItem, new[] { parentId, ribbonItem.Name });
+            return ribbonItem;
         }
 
         /// <summary>
-        /// GetRibbonItem
+        /// Enable / Disable Show Text
         /// </summary>
+        /// <typeparam name="TRibbonItem">RibbonItem</typeparam>
         /// <param name="ribbonItem"></param>
+        /// <param name="showText"></param>
         /// <returns></returns>
-        public static Autodesk.Windows.RibbonItem GetRibbonItem(this RibbonItem ribbonItem)
+        public static TRibbonItem SetShowText<TRibbonItem>(this TRibbonItem ribbonItem, bool showText = false) where TRibbonItem : RibbonItem
         {
-            var revitRibbonItem = RevitRibbonControl
-                .RibbonControl.findRibbonItemById(ribbonItem.GetId());
-            return revitRibbonItem;
+            if (ribbonItem.GetRibbonItem() != null)
+                ribbonItem.GetRibbonItem().ShowText = showText;
+            return ribbonItem;
         }
 
         /// <summary>
-        /// GetRibbonPanel
+        /// Enable / Disable Show Image
         /// </summary>
-        /// <param name="ribbonPanel"></param>
+        /// <typeparam name="TRibbonItem">RibbonItem</typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="showImage"></param>
         /// <returns></returns>
-        public static Autodesk.Windows.RibbonPanel GetRibbonPanel(this RibbonPanel ribbonPanel)
+        public static TRibbonItem SetShowImage<TRibbonItem>(this TRibbonItem ribbonItem, bool showImage = false) where TRibbonItem : RibbonItem
         {
-            var type = typeof(RibbonPanel);
-            var _ribbonPanel = type.GetField("m_RibbonPanel",
-                BindingFlags.Instance | BindingFlags.NonPublic)
-                ?.GetValue(ribbonPanel) as Autodesk.Windows.RibbonPanel;
-            return _ribbonPanel;
+            if (ribbonItem.GetRibbonItem() != null)
+                ribbonItem.GetRibbonItem().ShowImage = showImage;
+            return ribbonItem;
         }
 
         /// <summary>
-        /// GetRibbonTab
+        /// Set RibbonItemSize
         /// </summary>
-        /// <param name="ribbonPanel"></param>
+        /// <typeparam name="TRibbonItem">RibbonItem</typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="itemSize"></param>
         /// <returns></returns>
-        public static Autodesk.Windows.RibbonTab GetRibbonTab(this RibbonPanel ribbonPanel)
+        public static TRibbonItem SetItemSize<TRibbonItem>(this TRibbonItem ribbonItem, Autodesk.Windows.RibbonItemSize itemSize = Autodesk.Windows.RibbonItemSize.Large) where TRibbonItem : RibbonItem
         {
-            return ribbonPanel.GetRibbonPanel().Tab;
+            if (ribbonItem.GetRibbonItem() != null)
+                ribbonItem.GetRibbonItem().Size = itemSize;
+            return ribbonItem;
         }
 
         /// <summary>
-        /// GetRibbonTab
+        /// Set RibbonItem Text
         /// </summary>
-        /// <param name="tabId"></param>
+        /// <typeparam name="TRibbonItem">RibbonItem</typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="text"></param>
         /// <returns></returns>
-        public static Autodesk.Windows.RibbonTab GetRibbonTab(string tabId)
+        public static TRibbonItem SetText<TRibbonItem>(this TRibbonItem ribbonItem, string text = "") where TRibbonItem : RibbonItem
         {
-            var ribbon = Autodesk.Windows.ComponentManager.Ribbon;
-            foreach (Autodesk.Windows.RibbonTab tab in ribbon.Tabs)
-            {
-                if (tab.Id == tabId)
-                {
-                    return tab;
-                }
-            }
-            return null;
+            if (text == null)
+                return ribbonItem;
+
+            if (text.Trim() == string.Empty)
+                return ribbonItem.SetShowText(false);
+
+            ribbonItem.ItemText = text;
+            return ribbonItem;
         }
 
         /// <summary>
-        /// Get GetRibbonTabs
+        /// Set RibbonItem ToolTip
         /// </summary>
+        /// <typeparam name="TRibbonItem">RibbonItem</typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="toolTip"></param>
         /// <returns></returns>
-        public static IList<Autodesk.Windows.RibbonTab> GetRibbonTabs()
+        public static TRibbonItem SetToolTip<TRibbonItem>(this TRibbonItem ribbonItem, string toolTip) where TRibbonItem : RibbonItem
         {
-            var ribbon = Autodesk.Windows.ComponentManager.Ribbon;
-            return ribbon.Tabs;
+            if (toolTip != null)
+                ribbonItem.ToolTip = toolTip;
+
+            return ribbonItem;
         }
 
         /// <summary>
-        /// GetRibbonPanel
+        /// Set RibbonItem LongDescription
         /// </summary>
-        /// <param name="tabId"></param>
-        /// <param name="panelEndWithId"></param>
+        /// <typeparam name="TRibbonItem">RibbonItem</typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="longDescription"></param>
         /// <returns></returns>
-        public static Autodesk.Windows.RibbonPanel GetTabRibbonPanel(string tabId, string panelEndWithId)
+        public static TRibbonItem SetLongDescription<TRibbonItem>(this TRibbonItem ribbonItem, string longDescription) where TRibbonItem : RibbonItem
         {
-            var ribbon = Autodesk.Windows.ComponentManager.Ribbon;
-            foreach (Autodesk.Windows.RibbonTab tab in ribbon.Tabs)
-            {
-                if (tab.Id == tabId)
-                {
-                    foreach (Autodesk.Windows.RibbonPanel panel in tab.Panels)
-                    {
-                        if (panel.Source.Id.EndsWith(panelEndWithId))
-                        {
-                            return panel;
-                        }
-                    }
-                }
-            }
-            return null;
+            if (longDescription != null)
+                ribbonItem.LongDescription = longDescription;
+
+            return ribbonItem;
         }
 
         /// <summary>
-        /// GetRibbonPanels
+        /// Set RibbonItem ToolTipImage
         /// </summary>
-        /// <param name="tabId"></param>
+        /// <typeparam name="TRibbonItem">RibbonItem</typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="toolTipImage"></param>
         /// <returns></returns>
-        public static IList<Autodesk.Windows.RibbonPanel> GetRibbonPanels(string tabId)
+        public static TRibbonItem SetToolTipImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource toolTipImage) where TRibbonItem : RibbonItem
         {
-            var ribbon = Autodesk.Windows.ComponentManager.Ribbon;
-            var list = new List<Autodesk.Windows.RibbonPanel>();
-            foreach (Autodesk.Windows.RibbonTab tab in ribbon.Tabs)
-            {
-                if (tab.Id == tabId)
-                {
-                    return tab.Panels;
-                }
-            }
-            return list;
+            if (toolTipImage != null)
+                ribbonItem.ToolTipImage = toolTipImage;
+
+            return ribbonItem;
         }
-
-        /// <summary>
-        /// GetRibbonItems
-        /// </summary>
-        /// <param name="tabId"></param>
-        /// <param name="panelEndWithId"></param>
-        /// <returns></returns>
-        public static IList<Autodesk.Windows.RibbonItem> GetRibbonItems(string tabId, string panelEndWithId)
-        {
-            var ribbon = Autodesk.Windows.ComponentManager.Ribbon;
-            var list = new List<Autodesk.Windows.RibbonItem>();
-            foreach (Autodesk.Windows.RibbonTab tab in ribbon.Tabs)
-            {
-                if (tab.Id == tabId)
-                {
-                    foreach (Autodesk.Windows.RibbonPanel panel in tab.Panels)
-                    {
-                        if (panel.Source.Id.EndsWith(panelEndWithId))
-                        {
-                            return panel.Source.Items.ToList();
-                        }
-                    }
-                }
-            }
-            return list;
-        }
-
-        /// <summary>
-        /// MoveRibbonPanel
-        /// </summary>
-        /// <param name="tabId"></param>
-        /// <param name="panelEndWithId"></param>
-        /// <param name="newIndex"></param>
-        /// <returns></returns>
-        public static bool MoveRibbonPanel(string tabId, string panelEndWithId, int newIndex)
-        {
-            if (GetRibbonPanels(tabId) is Autodesk.Windows.RibbonTab ribbonTab)
-            {
-                if (newIndex < 0) newIndex = 0;
-                if (newIndex >= ribbonTab.Panels.Count) newIndex = ribbonTab.Panels.Count;
-                foreach (var panel in ribbonTab.Panels)
-                {
-                    if (panel.Source.Id.EndsWith(panelEndWithId))
-                    {
-                        var index = ribbonTab.Panels.IndexOf(panel);
-                        ribbonTab.Panels.Move(index, newIndex);
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-
-        #region Autodesk.Windows.RibbonTab
-
-        #region Order
-
-        /// <summary>
-        /// MoveRibbonPanel to Position
-        /// </summary>
-        /// <param name="ribbonPanel"></param>
-        /// <param name="newIndex"></param>
-        public static void MoveRibbonPanel(this Autodesk.Windows.RibbonPanel ribbonPanel, int newIndex = 0)
-        {
-            var ribbonTab = ribbonPanel.Tab;
-            var panels = ribbonTab.Panels;
-            var length = panels.Count;
-            for (int i = 0; i < length; i++)
-            {
-                if (i == newIndex) continue;
-                if (panels[i] == ribbonPanel)
-                {
-                    ribbonTab.Panels.Move(i, newIndex);
-                    return;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Set Order of Panels by Title
-        /// </summary>
-        /// <param name="ribbonTab"></param>
-        public static void SetOrderPanels(this Autodesk.Windows.RibbonTab ribbonTab)
-        {
-            var order = ribbonTab.Panels.OrderBy(e => e.Source.Title).ToList();
-            var length = order.Count;
-            if (length <= 1) return;
-            for (int i = 0; i < length; i++)
-            {
-                var o = order[i];
-                for (int j = i; j < length; j++)
-                {
-                    if (j == i) continue;
-                    if (o == ribbonTab.Panels[j])
-                    {
-                        ribbonTab.Panels.Move(j, i);
-                    }
-                }
-            }
-        }
-
         #endregion
 
+        #region Set RibbonButton
         /// <summary>
-        /// Remove Tab
+        /// Set RibbonButton Image
         /// </summary>
-        /// <param name="ribbonTab"></param>
-        public static void Remove(this Autodesk.Windows.RibbonTab ribbonTab)
+        /// <typeparam name="TRibbonItem"></typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="image"></param>
+        /// <returns></returns>
+        public static TRibbonItem SetImage<TRibbonItem>(this TRibbonItem ribbonItem, string image) where TRibbonItem : RibbonItem
         {
-            var ribbon = Autodesk.Windows.ComponentManager.Ribbon;
-            ribbon.Tabs.Remove(ribbonTab);
+            return ribbonItem.SetImage(image.GetBitmapSource());
         }
 
+        /// <summary>
+        /// Set RibbonButton LargeImage
+        /// </summary>
+        /// <typeparam name="TRibbonItem"></typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="largeImage"></param>
+        /// <returns></returns>
+        public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, string largeImage) where TRibbonItem : RibbonItem
+        {
+            return ribbonItem.SetLargeImage(largeImage.GetBitmapSource());
+        }
+
+        /// <summary>
+        /// Set RibbonButton Image
+        /// </summary>
+        /// <typeparam name="TRibbonItem">RibbonButton</typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="image"></param>
+        /// <returns></returns>
+        public static TRibbonItem SetImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource image) where TRibbonItem : RibbonItem
+        {
+            if (image != null)
+                if (ribbonItem is RibbonButton ribbonButton)
+                    ribbonButton.Image = image.GetBitmapFrame(16, (frame) => { ribbonButton.Image = frame; });
+
+            return ribbonItem;
+        }
+
+        /// <summary>
+        /// Set RibbonButton LargeImage
+        /// </summary>
+        /// <typeparam name="TRibbonItem">RibbonButton</typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <param name="largeImage"></param>
+        /// <returns></returns>
+        public static TRibbonItem SetLargeImage<TRibbonItem>(this TRibbonItem ribbonItem, ImageSource largeImage) where TRibbonItem : RibbonItem
+        {
+            if (largeImage != null)
+            {
+                if (ribbonItem is RibbonButton ribbonButton)
+                {
+                    ribbonButton.LargeImage = largeImage.GetBitmapFrame(32, (frame) => { ribbonButton.LargeImage = frame; });
+                    if (ribbonButton.Image == null)
+                        ribbonButton.SetImage(ribbonButton.LargeImage);
+                }
+            }
+
+            return ribbonItem;
+        }
+        #endregion
+
+        #region QuickAccessToolBar
+        /// <summary>
+        /// Add RibbonItem to QuickAccessToolBar
+        /// </summary>
+        /// <typeparam name="TRibbonItem"></typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <returns></returns>
+        public static TRibbonItem AddQuickAccessToolBar<TRibbonItem>(this TRibbonItem ribbonItem) where TRibbonItem : RibbonItem
+        {
+            ribbonItem.GetRibbonItem()?.AddQuickAccessToolBar();
+            return ribbonItem;
+        }
+
+        /// <summary>
+        /// Add RibbonItem to QuickAccessToolBar
+        /// </summary>
+        /// <typeparam name="TRibbonItem"></typeparam>
+        /// <param name="ribbonItem"></param>
+        /// <returns></returns>
+        public static TRibbonItem RemoveQuickAccessToolBar<TRibbonItem>(this TRibbonItem ribbonItem) where TRibbonItem : RibbonItem
+        {
+            ribbonItem.GetRibbonItem()?.RemoveQuickAccessToolBar();
+            return ribbonItem;
+        }
         #endregion
     }
 }