+++ CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.7] / 2022-02-08
+- Add Availability class
+- Update xml documetation
+- Fix AutoSelect base64 or uri
+- Add string on SetImage / SetLargeImage
+- 2022-02-07
+- Add QuickAccessToolBar on Autodesk.Revit.UI.RibbonItem
+- Add Pack UIFrameworkRes Reference
+- Remove RibbonItem from RibbonPanel - Include Split/Down/Stacked
+- Test Online Icons Scale
+- Test Scale
+- Test IsDownloading Icon
+- Icon Select Frame
+- Teste PushButtonData Set Proprieties
+- Change CreatePulldownButton to params
+- Change CreateSplitButton to params
+- Add RibbonHelpExtension
+- Add RibbonItemDataExtension
+- Clear Code - Remove not used stuff
+- Add RibbonUtilExtension
+- Add RibbonTabExtension
+- TODO: Add PushButtonData, RibbonItemData SetImage / Description / Othes : DONE
+- TODO: Add pack component BitmapSource : DONE
+- TODO: Add Icon selector size with BitmapDecoder : DONE
+- TODO: Add Remove RibbonItem From RibbonPanel : DONE
+- 2022-02-03
+- Add Remove RibbonItem on the Panel
+
 ## [0.0.6] / 2022-02-02
 - Special Icons8 class
 - Update Example Version
@@ -80,6 +108,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.0.8]: ../../compare/0.0.7...0.0.8
+[0.0.7]: ../../compare/0.0.6...0.0.7
+[0.0.6]: ../../compare/0.0.5...0.0.6
+[0.0.5]: ../../compare/0.0.4...0.0.5
 [0.0.4]: ../../compare/0.0.3...0.0.4
 [0.0.3]: ../../compare/0.0.2...0.0.3
 [0.0.2]: ../../compare/0.0.1...0.0.2